refactor: rename com.generic.ult to com.generic.plannr
