+++ Uni_life_tracker/app/src/main/AndroidManifest.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
           xmlns:tools="http://schemas.android.com/tools"
-          package="com.generic.ult">
+          package="com.generic.plannr">
 
     <application
             android:allowBackup="true"
@@ -10,18 +10,18 @@
             android:roundIcon="@drawable/ic_plannr_p_black"
             android:supportsRtl="true"
             android:theme="@style/Theme.Plannr">
-        <activity android:name=".SettingsActivity"> </activity>
-        <activity android:name=".SignUpActivity"> </activity>
-        <activity android:name=".AddEventActivity"> </activity>
-        <activity android:name=".AddExpensesActivity"> </activity>
-        <activity android:name=".ExpensesActivity"> </activity>
-        <activity android:name=".SchoolActivity"> </activity>
+        <activity android:name="com.generic.plannr.SettingsActivity"> </activity>
+        <activity android:name="com.generic.plannr.SignUpActivity"> </activity>
+        <activity android:name="com.generic.plannr.AddEventActivity"> </activity>
+        <activity android:name="com.generic.plannr.AddExpensesActivity"> </activity>
+        <activity android:name="com.generic.plannr.ExpensesActivity"> </activity>
+        <activity android:name="com.generic.plannr.SchoolActivity"> </activity>
         <activity
-                android:name=".MainPageActivity"
+                android:name="com.generic.plannr.MainPageActivity"
                 android:label="Main"
                 android:exported="true"> </activity>
         <activity
-                android:name=".WelcomeActivity"
+                android:name="com.generic.plannr.WelcomeActivity"
                 android:theme="@style/Theme.Plannr"
                 tools:ignore="AllowBackup"
                 android:exported="true"
