+++ Uni_life_tracker/app/src/main/java/com/generic/plannr/AddEventActivity.java
@@ -1,4 +1,4 @@
-package com.generic.ult;
+package com.generic.plannr;
 
 import android.annotation.SuppressLint;
 import android.app.DatePickerDialog;
