+++ Uni_life_tracker/app/src/main/java/com/generic/plannr/MainPageActivity.java
@@ -1,4 +1,4 @@
-package com.generic.ult;
+package com.generic.plannr;
 
 import Entities.Event;
 import android.app.Activity;
