+++ Uni_life_tracker/app/src/main/java/com/generic/plannr/SchoolActivity.java
@@ -1,4 +1,4 @@
-package com.generic.ult;
+package com.generic.plannr;
 import android.content.Intent;
 import androidx.appcompat.app.AppCompatActivity;
 import androidx.drawerlayout.widget.DrawerLayout;
