+++ Uni_life_tracker/app/src/test/java/com/generic/plannr/DeadlineTest.java
@@ -1,4 +1,4 @@
-package com.generic.ult;//import static org.junit.jupiter.api.Assertions.*;
+package com.generic.plannr;//import static org.junit.jupiter.api.Assertions.*;
 import Entities.Deadline;
 import org.junit.*;
 import static org.junit.Assert.*;
