+++ Uni_life_tracker/app/src/test/java/com/generic/plannr/UseCaseTests/EventDateComparatorTest.java
@@ -1,4 +1,4 @@
-package com.generic.ult.UseCaseTests;
+package com.generic.plannr.UseCaseTests;
 
 import org.junit.*;
