+++ Uni_life_tracker/app/src/test/java/com/generic/plannr/UseCaseTests/SortTodaysEventsTest.java
@@ -1,20 +1,17 @@
-package com.generic.ult.UseCaseTests;
+package com.generic.plannr.UseCaseTests;
 
 import org.junit.*;
 import static org.junit.Assert.*;
 
-import static UseCases.GetTodaysEvents.getTodaysEvents;
 import static UseCases.SortTodaysEvents.sortByDate;
 import static UseCases.SortTodaysEvents.sortByPriority;
 
-import java.lang.reflect.Array;
 import java.time.LocalDateTime;
 import java.util.ArrayList;
 import java.util.Arrays;
 
 import Entities.Event;
 import UseCases.EventDateComparator;
-import UseCases.GetTodaysEvents;
 import UseCases.UserManager;
 
 public class SortTodaysEventsTest {
